Fixes Send Email on Reports, by replacing multi-select with a textbox.
Originally this control used the multi-select validation, but somebody upgraded that control significantly to the point where it broke the specialized scenario that this control was using.  I have replaced it with a plain textbox just to get it working again, although this drops support for selecting a user from the validation.