#2138 Makes the security logic more strict about kicking out a user with an expired security group.
It's hard to see the change, when looking at the diff, but what this does is there was an if statement about validate security group and it was checking if the group was expired only when using the don't validate security group option.  Really it needs to always check if the group is expired, but only check on the inidividual security node when that option is used.